Modules

Writers

FastIce.Writers.write_h5 — Method

write_h5(arch, grid, path, fields)

Write output fields in HDF5 format to a file on path for global grid.

FastIce.Writers.write_xdmf — Function

write_xdmf(arch, grid, path, fields, h5_names, timesteps=Float64(0.0))

Write Xdmf metadata to path for corresponding h5_names and fields for global grid. Saving time-dependant data can be achieved upon passing a vector to h5_names and timesteps.

LevelSets

FastIce.LevelSets.compute_levelset_from_dem! — Function

compute_levelset_from_dem!(arch::Architecture, launch, Ψ::Field, dem::AbstractField, dem_grid2D::UniformGrid, grid::UniformGrid, R=LinearAlgebra.I)

Compute level sets from dem.

FastIce.LevelSets.invert_levelset! — Method

invert_levelset!(arch::Architecture, launch, Ψ::Field, grid::UniformGrid)

Invert level set Ψ to set what's below the surface as "inside".