Modules

Grids

FastIce.Grids.CartesianGrid — Type

Rectilinear grid with uniform spacing.

Examples

julia> using FastIce.Grids

julia> grid = CartesianGrid(origin=(0.0,0.0), extent=(1.0,1.0), size=(4,4))
2D 4×4 CartesianGrid{Float64}:
    x ∈ [0.0–1.0]; Δx = 0.25
    y ∈ [0.0–1.0]; Δy = 0.25

FastIce.Grids.CartesianGrid — Method

CartesianGrid(; origin::NTuple{N,T}, extent::NTuple{N,T}, size::NTuple{N,I})

Create a Cartesian grid with a specified origin (bottom-south-west corner in 3D), spatial extent, and number of grid cells.

FastIce.Grids.DiscreteAxis — Type

Discretised one-dimensional interval. Grid type CartesianGrid is defined as a tuple of DiscreteAxis's.

FastIce.Grids.axis — Method

axis(grid::CartesianGrid, dim::Integer)

Return a DiscreteAxis corresponding to the spatial dimension dim.

FastIce.Grids.coord — Method

coord(grid::CartesianGrid{N}, loc::NTuple{N,Location}, inds::NTuple{N})

Return a tuple of spatial coordinates of a grid point at location loc and indices inds.

For vertex-centered locations, first grid point is at the origin. For cell-centered locations, first grid point at half-spacing distance from the origin.

FastIce.Grids.extent — Method

extent(grid::CartesianGrid)

Return the spatial extent of a CartesianGrid as a tuple.

FastIce.Grids.origin — Method

origin(grid::CartesianGrid)

Return the origin of a CartesianGrid. The origin corresponds to bottom-south-west corner of the grid in 3D.

FastIce.Grids.spacing — Method

spacing(grid::CartesianGrid)

Return the spacing of a CartesianGrid as a tuple.

FastIce.Grids.Δ — Method

Δ(grid::CartesianGrid)

Return the spacing of a CartesianGrid as a tuple. Same as spacing.

Distributed

FastIce.Architectures.Architecture — Method

Architecture(backend::Backend, topology::CartesianTopology) where {N}

Create a distributed Architecture using backend and topology. For GPU backends, device will be selected automatically based on a process id within a node.

FastIce.Distributed.CartesianTopology — Type

CartesianTopology

Represents N-dimensional Cartesian topology of distributed processes.

FastIce.Distributed.CartesianTopology — Method

CartesianTopology(dims::NTuple{N,Int}; comm=MPI.COMM_WORLD) where {N}

Create an N-dimensional Cartesian topology with dimensions dims and using base MPI communicator comm (by default, comm=MPI.COMM_WORLD). If all entries in dims are not equal to 0, the product of dims should be equal to the total number of MPI processes MPI.Comm_size(comm). If any (or all) entries of dims are 0, the dimensions in the corresponding spatial directions will be picked automatically.

FastIce.Distributed.DistributedMPI — Type

Trait structure used as a type parameter to indicate that the Architecture is a distributed MPI Architecture.

FastIce.Distributed.ExchangeInfo — Type

ExchangeInfo

Structure containing the information to exchange halos of one side of an N-dimensional array.

FastIce.Distributed.ExchangeInfo — Method

ExchangeInfo(::Val{S}, ::Val{D}, field::Field) where {S,D}

Create ExchangeInfo to exchange halos on side S in the spatial direction D.

FastIce.BoundaryConditions.apply_boundary_conditions! — Method

apply_boundary_conditions!(::Val{S}, ::Val{D},
                           arch::Architecture,
                           grid::CartesianGrid,
                           fields::NTuple{N,Field},
                           exchange_infos::NTuple{N,ExchangeInfo}; async=true) where {S,D,N}

Perform a non-blocking MPI exchange for a set of fields.

FastIce.Distributed.cartesian_communicator — Method

cartesian_communicator(t::CartesianTopology)

MPI Cartesian communicator for the topology.

FastIce.Distributed.coordinates — Method

coordinates(t::CartesianTopology)

Coordinates of a current process within a Cartesian topology.

FastIce.Distributed.dimensions — Method

dimensions(t::CartesianTopology)

Dimensions of the topology as NTuple.

FastIce.Distributed.gather! — Method

gather!(arch::Architecture{DistributedMPI}, dst, src::Field; kwargs...)

Gather the interior of a field src into a global array dst.

FastIce.Distributed.gather! — Method

gather!(dst::Union{AbstractArray{T,N},Nothing}, src::AbstractArray{T,N}, comm::MPI.Comm; root=0) where {T,N}

Gather local array src into a global array dst. Size of the global array size(dst) should be equal to the product of the size of a local array size(src) and the dimensions of a Cartesian communicator comm. The array will be gathered on the process with id root (root=0 by default). Note that the memory for a global array should be allocated only on the process with id root, on other processes dst can be set to nothing.

FastIce.Distributed.global_grid_size — Method

global_grid_size(t::CartesianTopology, local_size)

Return the global size for a structured grid.

FastIce.Distributed.global_rank — Method

global_rank(t::CartesianTopology)

Global id of a process in a Cartesian topology.

FastIce.Distributed.global_size — Method

global_size(t::CartesianTopology)

Total number of processes withing the topology.

FastIce.Distributed.has_neighbor — Method

has_neighbor(t::CartesianTopology, dim, side)

Returns true if there a neighbor process in spatial direction dim on the side side, or false otherwise.

FastIce.Distributed.local_grid — Method

local_grid(g::CartesianGrid, t::CartesianTopology)

Return a CartesianGrid covering the subdomain which corresponds to the current process.

FastIce.Distributed.neighbor — Method

neighbor(t::CartesianTopology, dim, side)

Returns id of a neighbor process in spatial direction dim on the side side, if this neighbor exists, or MPI.PROC_NULL otherwise.

FastIce.Distributed.neighbors — Method

neighbors(t::CartesianTopology)

Neighbors of a current process.

Returns NTuple containing process ids of the two immediate neighbors in each spatial direction, or MPI.PROC_NULL if no neighbor on a corresponding side.

FastIce.Distributed.node_name — Method

node_name(t::CartesianTopology)

Name of a node according to MPI.Get_processor_name().

FastIce.Distributed.node_size — Method

node_size(t::CartesianTopology)

Number of processes sharing the same node.

FastIce.Distributed.shared_communicator — Method

shared_communicator(t::CartesianTopology)

MPI communicator for the processes sharing the same node.

FastIce.Distributed.shared_rank — Method

shared_rank(t::CartesianTopology)

Local id of a process within a single node. Can be used to set the GPU device.

Kernel launch

FastIce.KernelLaunch.launch! — Method

launch!(arch::Architecture, grid::CartesianGrid, kernel::Pair{K,Args}; <keyword arguments>) where {K,Args}

Launch a KernelAbstraction kernel on a grid using the backend from arch. Either worksize or location must be provided as keyword arguments.

Keyword Arguments

worksize: worksize of a kernel, i.e. how many grid points are included in each spatial direction.

location[=nothing]: compute worksize as a size of the grid at a specified location. If only one location is provided, e.g. location=Vertex(), then this location will be used for all spacial directions.

offset[=nothing]: index offset for all grid indices as a CartesianIndex.

expand[=nothing]: if provided, the worksize is increased by 2*expand, and offset is set to -expand, or combined with user-provided offset.

hide_boundaries[=nothing]: instance of HideBoundaries, that will be used to overlap boundary processing with computations at inner points of the domain.

outer_width[=nothing]: if hide_boundaries is specified, used to determine the decomposition of the domain into inner and outer regions.

boundary_conditions[=nothing]: a tuple of boundary condition batches for each side of every spatial direction.

async[=false]: if set to true, will return immediately, without waiting on the kernel execution to finish.

Writers

FastIce.Writers.write_h5 — Method

write_h5(arch::Architecture, grid::CartesianGrid, path, fields)

Write output fields in HDF5 format to a file on path.

FastIce.Writers.write_h5 — Method

write_h5(arch::Architecture{DistributedMPI}, grid::CartesianGrid, path, fields)

Write output fields in HDF5 format to a file on path for global grid on distributed arch.

FastIce.Writers.write_xdmf — Function

write_xdmf(arch::Architecture, grid::CartesianGrid, path, fields, h5_names, timesteps=Float64(0.0))

Write Xdmf metadata to path for corresponding h5_names and fields. Saving time-dependant data can be achieved upon passing a vector to h5_names and timesteps.

FastIce.Writers.write_xdmf — Function

write_xdmf(arch::Architecture{DistributedMPI}, grid::CartesianGrid, path, fields, h5_names, timesteps=Float64(0.0))

Write Xdmf metadata to path for corresponding h5_names and fields for global grid on distributed arch. Saving time-dependant data can be achieved upon passing a vector to h5_names and timesteps.

LevelSets

FastIce.LevelSets._init_level_set! — Method

_init_level_set!(Ψ::Field, dem::Field, dem_grid::CartesianGrid, Ψ_grid::CartesianGrid, cutoff::AbstractFloat, R::AbstractMatrix)

Initialize level sets.

FastIce.LevelSets.compute_level_set_from_dem! — Function

compute_level_set_from_dem!(arch::Architecture, Ψ::Field, dem::Field, dem_grid::CartesianGrid, Ψ_grid::CartesianGrid, R=LinearAlgebra.I)

Compute level sets from dem.